+++ CRM/Trackpaypal/Form/Settings.php
@@ -13,6 +13,7 @@ class CRM_Trackpaypal_Form_Settings extends CRM_Core_Form {
     return array(
       'standard' => 'standard',
       'ecommerce' => 'ecommerce',
+      'both' => 'both',
     );
   }