+++ CHANGELOG.md
@@ -6,13 +6,16 @@ The extension is licensed under [AGPL-3.0](LICENSE.txt).
 
 ## Unreleased
 
-* Push ecommerce event to Google Analytics
+* Fixed typo in variable name
+* New option to send ecommerce transaction _and_ event, alongside one or the other
+* Simpler language for settings page field labels
 
 ## 0.1.0 - 2018-11-06
 
 ### Added
 
 * CHANGELOG.md
+* Push ecommerce event or standard event to Google Analytics
 
 ### Changed